requirements.txt
Django
gunicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-exporter-otlp
python-dotenv
